+++ FirebaseAI/Sources/Types/Public/Live/LiveGenerationConfig.swift
@@ -20,6 +20,7 @@ public struct LiveGenerationConfig: Sendable {
   let bidiGenerationConfig: BidiGenerationConfig
   let inputAudioTranscription: BidiAudioTranscriptionConfig?
   let outputAudioTranscription: BidiAudioTranscriptionConfig?
+  let contextWindowCompression: BidiContextWindowCompressionConfig?
 
   /// Creates a new ``LiveGenerationConfig`` value.
   ///
@@ -118,13 +119,18 @@ public struct LiveGenerationConfig: Sendable {
   ///
   ///     > Important: Transcripts are independent to the model turn. This means transcripts may
   ///     > come earlier or later than when the model sends the corresponding audio responses.
+  ///   - contextWindowCompression: Enables context window compression to manage the model's
+  ///     context window.
+  ///
+  ///     This mechanism prevents the context from exceeding a given length.
   public init(temperature: Float? = nil, topP: Float? = nil, topK: Int? = nil,
               candidateCount: Int? = nil, maxOutputTokens: Int? = nil,
               presencePenalty: Float? = nil, frequencyPenalty: Float? = nil,
               responseModalities: [ResponseModality]? = nil,
               speech: SpeechConfig? = nil,
               inputAudioTranscription: AudioTranscriptionConfig? = nil,
-              outputAudioTranscription: AudioTranscriptionConfig? = nil) {
+              outputAudioTranscription: AudioTranscriptionConfig? = nil,
+              contextWindowCompression: ContextWindowCompressionConfig? = nil) {
     self.init(
       BidiGenerationConfig(
         temperature: temperature,
@@ -138,15 +144,18 @@ public struct LiveGenerationConfig: Sendable {
         speechConfig: speech?.speechConfig
       ),
       inputAudioTranscription: inputAudioTranscription?.audioTranscriptionConfig,
-      outputAudioTranscription: outputAudioTranscription?.audioTranscriptionConfig
+      outputAudioTranscription: outputAudioTranscription?.audioTranscriptionConfig,
+      contextWindowCompression: contextWindowCompression?.bidiContextWindowCompressionConfig
     )
   }
 
   init(_ bidiGenerationConfig: BidiGenerationConfig,
        inputAudioTranscription: BidiAudioTranscriptionConfig? = nil,
-       outputAudioTranscription: BidiAudioTranscriptionConfig? = nil) {
+       outputAudioTranscription: BidiAudioTranscriptionConfig? = nil,
+       contextWindowCompression: BidiContextWindowCompressionConfig? = nil) {
     self.bidiGenerationConfig = bidiGenerationConfig
     self.inputAudioTranscription = inputAudioTranscription
     self.outputAudioTranscription = outputAudioTranscription
+    self.contextWindowCompression = contextWindowCompression
   }
 }